move AddJournal
