+++ src/components/index.js
@@ -1,1 +1,3 @@
 export * from './marginals';
+
+export { default as AddJournal} from './AddJournal/AddJournal'
