+++ src/components/marginals/index.js
@@ -1,2 +1,3 @@
+
 export { default as Footer } from "./Footer/Footer";
 export {default as Navbar } from "./Navbar/Navigation"
